media: hantro: add support for RK3188
RK3188s VPU IP-Block is the predecessor from what RK3288 has.

While most of the registers match in the current state of the driver
there are some HW differences:
 - supports resultion up to 1920x1088 only
 - has one aclk and one hclk per vdpu/vepu
 - doesn't have the 'G1_REG_SOFT_RESET' register
 - ACLKs can be clocked up to 300 MHz only
 - no MMU for VPU
These make it necessary to add another variant to the driver.

This should be also reuseable for RK3066 but has not been tested,
so I'm not adding it here.