media: hantro: add support for Rockchip RK3036
RK3036 shares the hantro VPU IP Block with RK3288.
HW differences are:
- supports decoding up to 1920x1088 only
- ACLK can be clocked at max. 300MHz
This makes adding another variant to the driver necessary here also.

However RK3036 TRM does not mention that it has an encoder also. I verfied
this on my devices and it worked for JPEG encoding. Since the identification
register share the id with RK3288 and the fact that it has an interrupt for
video encoding, makes me think that the encoding IP is the same.

This variant could also be used for RK312x, but has not been tested.